# packages/skyhelper/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{user}": {
      "get": {
        "description": "Retrieves user settings and preferences",
        "operationId": "UsersController_getUser",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "description": "Discord user ID",
            "schema": {
              "example": "123456789012345678"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "language": {
                      "type": "string",
                      "example": "en-US"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user information",
        "tags": ["User Management"]
      },
      "patch": {
        "description": "Updates user preferences like language",
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "description": "Discord user ID",
            "schema": {
              "example": "123456789012345678"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "language": {
                    "type": "string",
                    "example": "en-US"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "language": {
                      "type": "string",
                      "example": "en-US"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user settings",
        "tags": ["User Management"]
      }
    },
    "/users/{user}/linked-role": {
      "patch": {
        "description": "Updates Discord linked role connection metadata for the user",
        "operationId": "UsersController_updateUserRoleMetadata",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "description": "Discord user ID",
            "schema": {
              "example": "123456789012345678"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Role metadata to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Platform username"
                  },
                  "metadata": {
                    "type": "object",
                    "properties": {
                      "wings": {
                        "type": "number",
                        "minimum": 1,
                        "maximum": 240,
                        "description": "Number of wings"
                      },
                      "since": {
                        "type": "string",
                        "description": "Date since playing"
                      },
                      "eden": {
                        "type": "boolean",
                        "description": "Has completed Eden"
                      },
                      "cr": {
                        "type": "boolean",
                        "description": "Completed candle runs"
                      },
                      "hangout": {
                        "type": "boolean",
                        "description": "Available for hangouts"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role metadata updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user linked role metadata",
        "tags": ["User Management"]
      },
      "get": {
        "description": "Retrieves Discord linked role connection metadata for the user",
        "operationId": "UsersController_getUserRoleMetadata",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "description": "Discord user ID",
            "schema": {
              "example": "123456789012345678"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role metadata retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "wings": {
                          "type": "number"
                        },
                        "since": {
                          "type": "string"
                        },
                        "eden": {
                          "type": "boolean"
                        },
                        "cr": {
                          "type": "boolean"
                        },
                        "hangout": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user linked role metadata",
        "tags": ["User Management"]
      }
    },
...
  },
  "info": {
    "title": "SkyHelper API",
    "description": "\n      The SkyHelper Bot API provides comprehensive access to Sky: Children of the Light game data and bot management features.\n      \n      ## Features\n      - **Guild Management**: Configure bot settings for Discord servers\n      - **Game Data**: Access shards, events, traveling spirits, and daily quests information\n      - **User Preferences**: Manage user settings and linked roles\n      - **Live Updates**: Configure automated updates for game events\n      - **Reminders**: Set up notifications for game activities\n      \n      ## Authentication\n      Most endpoints require authentication via Bearer token obtained through Discord OAuth2.\n      The token should be included in the Authorization header as: `Bearer <token>`\n      \n      ## Rate Limiting\n      API requests are rate-limited to prevent abuse. Please respect the limits and implement proper retry logic.\n      \n      ## Support\n      For API support, join our Discord server or check the GitHub repository.\n    ",
    "version": "7.7.0",
    "contact": {
      "name": "SkyHelper Team",
      "url": "https://github.com/imnaiyar/SkyHelper",
      "email": "support@skyhelperbot.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/imnaiyar/SkyHelper/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://api.skyhelperbot.com",
...
    },
    {
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your Discord OAuth2 access token",
        "in": "header"
      }
    },
...
  }
}